(Saimon) Implemented minimal working flow

+++ src/main/java/com/great/cms/db/dao/UserDao.java
@@ -12,7 +12,7 @@ import com.great.cms.db.entity.User;
  * @author ziniapc
  *
  */
-public interface UserDao extends GenericDao<User, Long> {
+public interface UserDao extends GenericDao<User, Integer> {
