+++ src/main/java/com/great/cms/service/UserService.java
@@ -4,6 +4,6 @@ import com.great.cms.db.entity.User;
 
 public interface UserService {
 
-	public User getUserByID(long id);
+	public User getUserByID(Integer id);
 	public User getUserByName(String userName);
 }
